requirements.txt
Flask
SQLAlchemy
APScheduler
fpdf2
requests
gunicorn
